# Faster exp calculation